requirements.txt
amqp==2.6.0
asgiref==3.2.10
attrs==19.3.0
Automat==20.2.0
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
cssselect==1.1.0
Django==3.0.8
django-3-jet==1.0.8
django-cors-headers==3.4.0
django-filter==2.0.0
django-filters==0.2.1
django-sendsms==0.3.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docopt==0.6.2
drf-user==0.0.8
drf-yasg==1.17.1
drfaddons==0.1.3
elasticsearch==7.8.0
flex==6.14.1
flower==0.9.5
future==0.18.2
humanize==2.5.0
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
inflection==0.5.0
itemadapter==0.1.0
itypes==1.2.0
Jinja2==2.11.2
jsonpointer==2.0
jsonschema==3.2.0
kombu==4.6.11
lxml==4.5.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.4
parsel==1.6.0
prometheus-client==0.8.0
Protego==0.1.16
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
pyspark==3.0.0
pyspark-stubs==3.0.0.dev8
pytube3==9.6.4
pytz==2020.1
PyYAML==5.3.1
queuelib==1.5.0
redis==3.5.3
requests==2.24.0
rfc3987==1.3.8
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
Scrapy==2.2.0
service-identity==18.1.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
strict-rfc3339==0.7
swagger-spec-validator==2.7.3
tornado==6.0.4
Twisted==20.3.0
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.9
uuid==1.30
validate-email==1.3
vine==1.3.0
w3lib==1.22.0
webvtt-py==0.4.5
youtube-transcript-api==0.3.1
zope.interface==5.1.0
